% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_into.R
\name{group_into}
\alias{group_into}
\title{Adding a new categorical column into the table}
\usage{
group_into(table, source_name, new_name, categories)
}
\arguments{
\item{table}{original table need to be modified (it makes sense to name the}

\item{source_name}{name of the column to be categorized (the name is intuitive)}

\item{new_name}{name of the new column which has categorized values (the name}

\item{categories}{a vector of categories, it needs to be in ascending order}
}
\value{
the table with a new categorical column
}
\description{
This function will create a new table that includes the original table and a
new column which is a categorical variable with values depending on the values
requested by the user.
}
\examples{
new_cancer_sample <- group_into(cancer_sample, "area_mean", "area", c("Small", "Medium", "Large"))
}
